+++ Changelog.md
@@ -2,6 +2,8 @@
 
 - New: ProjectE support (all reinforced blocks now have an EMC value)
 - New: Taser effects and damage are now configurable
+- New: The Block Change Detector can now show highlights in the world, marking where it detected block changes
+- New: Ability to change a block pocket's outline color
 - Fix: Nether portals can replace reinforced blocks when generating
 - Fix: Can't place blocks on blocks locked by a Sonic Security System
 - Fix: Sentry bullets can remove item frames